feat: get all the groups which have Chandigarh's coords in address
Phew, rust is fun. Gotta refactor the code to not be a total mess, abstract some
functions to make them reusable; and proceed to saving the groups in sqlite.
From there we'll proceed to writing code for fetching events and members of the
groups based on when their events/members were last retrieved; or perhaps we can
get some kind of indicator that a group's events or members have changed. Rough
ideas.