feat: first iteration MVP of landing page
It just has the list of events for now, switching to initiative and RFCs isn't
possible. I think it's better if I integrate this with Rust and get a list of
upcoming events rolling first. Then we'll launch the site as a simple
github-page, after which we'll pull initiatives and RFCs.