feat: reasonable event details page
It is not too fancy, but it conveys the information we have well enough. From
here we can proceed to build a static site which can be deployed on github
pages. There are a lot of features still to be implemented for Entropy; this is
incomplete, but fastest way to ship