feat: rfcs page populated from the rfcs directory
I have decided to make markdown the only supported format for now. Org-mode is
super-cool, but super unsupported. Also quite alien to most people.

I am putting a pin in this at this point, and am going to add a layer of
abstraction/pre-processing here. Need to read meta-data from the markdown files,
and augment the database using md files. e.g it will enable overriding an event's
information. But for starters, we'll take only RFCs and Initiatives.