refactor: part ways with async poaching meetup
I need better understanding of how to properly utilize async in Rust. I've
removed spawning a new task for every entry in meetup poacher, and have
essentially made it sync now. Result is easier to follow code, which closes
sender/receivers on completion. But it work on each poacher entry one at a time
in series.

Sometime in future when I understand async in rust a bit better, I'll come back
to this and make poacher async again.