feat: get correct rfc/initiative count on home page from scraped initiatives
For now, we are treating any initiative without any goals as an "RFC". Need to
think more on it. I am considering adding a "draft" state to goals; so any
initiative with no published goals is considered an RFC. We also need to decide
how to discuss RFCs, and how to publish those discussion on the RFC page on
entropy. I am thinking of using github issues page for now, but need to put more
thought into it.