chore: new meetup events
One thing lead to another and we are doing a meetup this week. Instead of
sorting everything up, I am creating the event on HEAD, pushing it to gh-pages,
and calling it a day. In this meetup itself, I'll pitch for contributors for
Entropy itself. Let's see how this goes.