feat: post-process built HTML to make storage assets available locally
Since we aren't hosting the site anywhere anymore, building locally and
deploying to gh-pages is how it is. But storage assets, which need minio
installed and available break in that process.

Instead of hosting the app somewhere, I've decided to download the storage
assets and make them available to the static site as static assets. I believe
this is the right thing to do since keeping the operational costs of an entropy
deployment as small as possible is a project goal.